Changed plugin ID's to unsigned long's, moved stuff from view menu to tools menu and removed view menu, removed dmx address allocation system because it's pretty useless and causes trouble, plus some messagebox improvements.

+++ libs/common/plugin.cpp
@@ -21,7 +21,7 @@
 
 #include "plugin.h"
 
-Plugin::Plugin(int id) : QObject()
+Plugin::Plugin(unsigned long id) : QObject()
 {
   m_id = id;
 }
